requirements.txt
numpy
pillow
opencv-python
scaleapi
requests